パース処理をIndexedDBからの読み取り時からIndexedDBへの登録時に変更
