+++ src/data/usePayments.ts
@@ -1,24 +1,9 @@
-import { useEffect, useState } from "react";
 import { useLiveQuery } from "dexie-react-hooks";
-import { parse } from "csv-parse/browser/esm/sync";
-import { z } from "zod";
 import { db } from "./db";
-import Encoding from "encoding-japanese";
 
 type Props = {
   fileName: string;
 };
-
-const paymentSchema = z.object({
-  date: z.string(),
-  name: z.string(),
-  price: z.string(),
-  count: z.string(),
-});
-
-type Payment = z.infer<typeof paymentSchema>;
-
-type _Payment = Omit<Payment, "price"> & { price: number };
 
 export function usePayments({ fileName }: Props) {
   const files = useLiveQuery(() =>
@@ -26,63 +11,9 @@ export function usePayments({ fileName }: Props) {
   );
   const file = files ? files[0] : undefined;
 
-  const [payments, setPayments] = useState<_Payment[]>();
-
-  useEffect(() => {
-    void (async () => {
-      if (!file) {
-        return [];
-      }
-
-      const arrayBuffer = await file.blob.arrayBuffer();
-      const uint8 = new Uint8Array(arrayBuffer);
-      const paymentText = Encoding.convert(uint8, {
-        from: "SJIS",
-        to: "UTF8",
-      });
-
-      const records = parse(paymentText, {
-        columns: [
-          "date",
-          "name",
-          "price",
-          "count",
-          "noname1",
-          "noname2",
-          "noname3",
-        ],
-        skip_empty_lines: true,
-        skipRecordsWithError: true,
-      });
-
-      const _payments = records.map((record) => {
-        const payment = paymentSchema.safeParse(record);
-        if (payment.success) {
-          return payment.data;
-        } else {
-          console.error("Invalid record:", record, payment.error);
-          throw new Error("Invalid payment record");
-        }
-      });
-
-      _payments.sort((a, b) => {
-        return new Date(a.date).getTime() - new Date(b.date).getTime();
-      });
-
-      setPayments(
-        _payments
-          .map((payment) => ({
-            ...payment,
-            price: parseInt(payment.price, 10),
-          }))
-          .filter((payment) => payment.name.length > 0),
-      );
-    })();
-  }, [file]);
-
-  if (!payments) {
+  if (!file) {
     return { status: "loading" as const };
   }
 
-  return { status: "completed" as const, payments };
+  return { status: "completed" as const, payments: file.payments };
 }
